# store/internal-api-openapi.yaml
openapi: 3.0.1
info:
  title: Carbonio Mailbox Internal API
  description: Service-to-service REST API for carbonio-mailbox
  version: "1.0"
paths:
  /accounts/batch:
    post:
      summary: Batch Get Accounts
      description: Returns account info for a list of IDs or emails. Exactly one of
        'ids' or 'emails' must be provided. Unknown entries are silently skipped.
      operationId: batchGetAccounts
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchRequest"
      responses:
        "200":
          description: List of Account Info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfoResponse"
        "400":
          description: "Invalid request: exactly one of 'ids' or 'emails' required,\
            \ max 100 entries"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/count-by-cos:
    get:
      summary: Count Accounts by COS
      description: "Counts the accounts in each of the given COSes. Uses the same\
        \ per-domain tally the license counts with, so an account inheriting the domain\
        \ or default COS is attributed to that COS."
      operationId: countAccountsByCos
      parameters:
      - name: cosId
        in: query
        description: "The COS ids to count, repeated"
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Accounts per COS and their sum
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountCountByCosResponse"
        "400":
          description: "Missing cosId query parameter, or too many"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts:
    get:
      summary: Get Account by Email
      description: Returns account info by email address
      operationId: getAccountByEmail
      parameters:
      - name: email
        in: query
        description: The account email address
        schema:
          type: string
      responses:
        "200":
          description: Account Info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfoResponse"
        "400":
          description: Missing email query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/info:
    get:
      summary: Get Account Info
      description: Returns account info by ID
      operationId: getAccountInfo
      parameters:
      - name: id
        in: path
        description: The account ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Account Info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfoResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/myself:
    get:
      summary: Get Authenticated Account Info
      description: Returns account info for the authenticated user
      operationId: getMyAccountInfo
      parameters:
      - name: ZM_AUTH_TOKEN
        in: cookie
        schema:
          type: string
      - name: ZM_ADMIN_AUTH_TOKEN
        in: cookie
        schema:
          type: string
      responses:
        "200":
          description: Account Info
        "401":
          description: Invalid or missing auth token
        "404":
          description: Account not found
        "500":
          description: Internal server error
  /accounts/{id}/shared-accounts:
    get:
      summary: Get Shared Accounts
      description: Returns accounts that share folders with the given account (via
        mountpoints)
      operationId: getSharedAccounts
      parameters:
      - name: id
        in: path
        description: The account ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: List of shared accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedAccountResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cos/{id}:
    get:
      summary: Get COS
      description: "Returns a class of service by ID. A 404 means no COS carries that\
        \ ID, which callers can read as non-existence."
      operationId: getCos
      parameters:
      - name: id
        in: path
        description: The COS ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: COS info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CosInfoResponse"
        "404":
          description: COS not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/mail/usage/{accountId}:
    get:
      summary: Get Mailbox usage
      description: Returns mail usage for the account mailbox
      operationId: getMailUsage
      parameters:
      - name: accountId
        in: path
        description: The account ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Mailbox usage in bytes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MailUsageResponse"
        "404":
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    AccountInfoResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        displayName:
          type: string
        cosId:
          type: string
        domainId:
          type: string
        domain:
          type: string
        status:
          type: string
          enum:
          - active
          - maintenance
          - locked
          - closed
          - lockout
          - pending
        isGlobalAdmin:
          type: boolean
        isExternal:
          type: boolean
        locale:
          type: string
        features:
          type: object
          additionalProperties:
            type: boolean
        capabilities:
          type: object
          additionalProperties:
            type: string
        sessionLifetimeMs:
          type: integer
          format: int64
        isExternalVirtualAccount:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    BatchRequest:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
        emails:
          type: array
          items:
            type: string
    AccountCountByCosResponse:
      type: object
      properties:
        total:
          type: integer
          format: int64
        byCos:
          type: object
          additionalProperties:
            type: integer
            format: int64
    SharedAccountResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        domainId:
          type: string
        cosId:
          type: string
    CosInfoResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    MailUsageResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        used:
          type: integer
          format: int64